refactor: update import paths for actions and add server directive to enhance code organization and clarity
