+++ lib/utils/supabase/server.ts
@@ -1,3 +1,5 @@
+'use server';
+
 import { createServerClient } from '@supabase/ssr'
 import { cookies } from 'next/headers'
